@@ -1,8 +1,9 @@
 ﻿using BHD_ServerManager.Forms;
-using BHD_ServerManager.Classes.CoreObjects;
+using HawkSyncShared;
+using HawkSyncShared.SupportClasses;
 using BHD_ServerManager.Classes.GameManagement;
 using BHD_ServerManager.Classes.InstanceManagers;
-using BHD_ServerManager.Classes.Instances;
+using HawkSyncShared.Instances;
 using BHD_ServerManager.Classes.SupportClasses;
 using System.Net;
 using System.Text;
