@@ -1,8 +1,9 @@
 ﻿using BHD_ServerManager.Classes.GameManagement;
 using BHD_ServerManager.Forms;
-using BHD_ServerManager.Classes.CoreObjects;
-using BHD_ServerManager.Classes.Instances;
-using BHD_ServerManager.Classes.ObjectClasses;
+using HawkSyncShared;
+using HawkSyncShared.SupportClasses;
+using HawkSyncShared.Instances;
+using HawkSyncShared.ObjectClasses;
 using BHD_ServerManager.Classes.SupportClasses;
 using System.Net.Http.Headers;
 using System.Text;
