@@ -1,6 +1,7 @@
-﻿using BHD_ServerManager.Classes.CoreObjects;
+﻿using HawkSyncShared;
+using HawkSyncShared.SupportClasses;
 using BHD_ServerManager.Classes.InstanceManagers;
-using BHD_ServerManager.Classes.Instances;
+using HawkSyncShared.Instances;
 using BHD_ServerManager.Classes.SupportClasses;
 using System;
 using System.Collections.Generic;
