@@ -1,7 +1,7 @@
-﻿using BHD_ServerManager.Classes.ObjectClasses;
+﻿using HawkSyncShared.ObjectClasses;
 using System.Collections.Generic;
 
-namespace BHD_ServerManager.Classes.Instances;
+namespace HawkSyncShared.Instances;
 
 /// <summary>
 /// Contains all player-related data that changes frequently during gameplay
