@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace BHD_ServerManager.Classes.ObjectClasses
+namespace HawkSyncShared.ObjectClasses
 {
     public class objectGameTypes
     {
